+++ src/app/components/flows/flowDesigner/flowDesignerPageController.js
@@ -3,13 +3,13 @@
 angular.module('liveopsConfigPanel')
   .controller('DesignerPageController', ['$scope', '$window', '$state', '$sce', 'Session', 'UserPermissions', 'apiHostname', 'designerHostname',
     function($scope, $window, $state, $sce, Session, UserPermissions, apiHostname, designerHostname) {
-    
+      /* globals window */
+
       $scope.designerHostname = $sce.trustAsResourceUrl(designerHostname);
 
       function ProcessMessage(event){
         switch(event.data.message){
           case 'FlowDesigner.ready':
-            console.log('Flow Designer is ready');
             event.source.postMessage({
               message: 'FlowDesigner.start',
               data: {
@@ -20,18 +20,18 @@ angular.module('liveopsConfigPanel')
                 flowId: $state.params.flowId,
                 draftId: $state.params.draftId
               }
-            }, '*')
+            }, '*');
             break;
           case 'FlowDesigner.versionPublished':
             $state.go('content.flows.flowManagement', {}, {reload: true});
             break;
         }
       }
-    
+
       window.addEventListener('message', ProcessMessage);
 
       $scope.$on('$destroy', function(){
-        window.removeEventListener('message', ProcessMessage)
-      })
+        window.removeEventListener('message', ProcessMessage);
+      });
     }
   ]);